MeshObjects::TopologicalMeshObject: Replaced by MoveableMeshObject with a dummy movePoints function
It is simpler and more consistent to declare topological mesh objects as
moveable with a dummy movePoints function so that they remain unchanged if the
mesh moves without topology change.  This allows the TopologicalMeshObject to be
removed altogether, the base MeshObjects type now being DeletableMeshObject
which are deleted on mesh change.