fieldMapper: Specificiation of unmapped values
Mapping functions can now, where appropriate, tell a field mapper to
assign specific values for un-mapped elements of the field. This is
achieved by means of new operators which take an additional argument of
either a value or a field-generating function. This value or function is
used to initialise the unmapped elements.