Time: Adjust the precision of the time name if necessary every time-step
This change guarantees that the precision of the time written to the log file,
all functionObject logs and the time directories is sufficient to be unique
irrespective of the change in the time-step or start time of the run.  If the
calculated time precision exceeds what can be represented by the floating point
resolution, e.g. by the time-step dropping to a VERY low value, the run stops.