New lumpedMassTemperature boundary condition
This boundary condition is applied to a patch which bounds a solid body,
wholly or partially.  It represents the body as a lumped mass, i.e. by a
single temperature which is fixed across the patch.  The body has a
volume which is either specified by the user, or is calculated when the
patch describes a closed volume (including in 2D meshes). Starting from
an initial temperature, the change in temperature is the calculated over
time according to a specified applied power source and the heat
transferred across the boundary, using a specified density and specific
heat capacity of the lumped mass.