IOerror: Print the error message only from the master if the dictionary is global
This avoids error messages from all processors if the dictionary being parsed is
global, in which case the error message from all processors would be the same so
only the master need print it.  This makes many common error messages the same
running in serial and parallel for user convenience.